python_basics_course/08_auth_payment_nextjs/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.6.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-decouple==3.8
pydantic[email]==2.5.2
requests==2.31.0
